+++ src/main/java/com/advancia/stage/dao/IngredienteDAO.java
@@ -1,7 +1,6 @@
 package com.advancia.stage.dao;
 
 import com.advancia.stage.dto.IngredienteDTO;
-import com.advancia.stage.model.Impasto;
 import com.advancia.stage.model.Ingrediente;
 import com.advancia.stage.util.ListModel;
 import com.advancia.stage.util.SingleModel;
@@ -20,40 +19,59 @@ public class IngredienteDAO {
     EntityManager em;
 
     public List<IngredienteDTO> ing_list(){
-        List<IngredienteDTO> list = new ArrayList<IngredienteDTO>();
+
 
         Query query = em.createQuery("from Ingrediente");
         List<Ingrediente> res = (ArrayList<Ingrediente>) query.getResultList();
-        list = ListModel.transformToIngredienteDTO(res);
-        return list;
+        if(res != null) {
+            return ListModel.transformToIngredienteDTO(res);
+        }
+        return null;
     }
 
     public IngredienteDTO getIngrediente(Long id){
-        Query query = em.createQuery("from Ingrediente i where i.id = "+id);
-        List<Ingrediente> x = query.getResultList();
-        return SingleModel.convertToIngredienteDTO(x.get(0));
+        if(id != null && id != 0) {
+            Query query = em.createQuery("from Ingrediente i where i.id = " + id);
+            Ingrediente x = (Ingrediente) query.getSingleResult();
+            if(x != null) {
+                return SingleModel.convertToIngredienteDTO(x);
+            }
+        }
+        return null;
     }
 
     public void add(Ingrediente ingrediente){
-        em.persist(ingrediente);
-        System.out.println("inserimento ingrediente");
-        em.flush();
-        em.clear();
+        if(ingrediente != null) {
+            em.persist(ingrediente);
+            System.out.println("inserimento ingrediente");
+            em.flush();
+            em.clear();
+        }
     }
 
     public void delete(Long id){
-        Ingrediente ingrediente = em.find(Ingrediente.class, id);
-        em.remove(ingrediente);
-        em.flush();
-        em.clear();
+        if(id != null && id != 0) {
+            Ingrediente ingrediente = em.find(Ingrediente.class, id);
+            if(ingrediente != null) {
+                em.remove(ingrediente);
+                em.flush();
+                em.clear();
+            }
+        }
     }
 
     public void update(Long id, Ingrediente ingrediente){
-        Ingrediente ingredienteVecchio =em.find(Ingrediente.class, id);
-        if(ingrediente.getNome() != null){
-            ingredienteVecchio.setNome(ingrediente.getNome());
-        }
+        if(id != null && id != 0) {
+            if(ingrediente != null) {
+                Ingrediente ingredienteVecchio = em.find(Ingrediente.class, id);
+                if(ingredienteVecchio != null) {
+                    if (ingrediente.getNome() != null) {
+                        ingredienteVecchio.setNome(ingrediente.getNome());
+                    }
 
-        em.merge(ingredienteVecchio);
+                    em.merge(ingredienteVecchio);
+                }
+            }
+        }
     }
 }
